test: test config
